rpi5_yolo_whisper/requirements_rpi5.txt
# Raspberry Pi 5 - YOLO + Whisper Requirements
# Python 3.11+

# Core dependencies
numpy>=1.24.0
python-dotenv>=1.0.0

# YOLO Object Detection
ultralytics>=8.0.0
opencv-python>=4.8.0

# Whisper Speech Recognition (faster-whisper for Pi optimization)
faster-whisper>=1.0.0
openai-whisper>=20231117

# Audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.13

# Text-to-Speech (offline)
pyttsx3>=2.90

# NCNN for optimized YOLO inference on Raspberry Pi
ncnn>=1.0.0

# ONNX for model conversion (intermediate step)
onnx>=1.14.0
onnxruntime>=1.15.0

# Utilities
tqdm>=4.66.0

# GUI Support
pillow>=10.0.0
